Desenvolvimento/Servidor/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-tastypie==0.14.2
djangorestframework==3.9.1
geographiclib==1.49
geopy==1.18.1
mysqlclient==1.4.0
PyMySQL==0.9.3
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.9
requirements==0.1
six==1.12.0
gunicorn==19.7.1
virtualenv==16.1.0